# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "API Docs for Inventory App",
    "version": "1.0",
    "title": "Inventory API"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "localhost:8080/api/v1"
    }
  ],
  "paths": {
...
    "/locations": {
      "post": {
        "summary": "Create a Location",
        "description": "Create a new Location for the requesting user. Users with free accounts can create up to one Location (either a fridge or a pantry). Users with premium accounts can create unlimited Locations",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationCreateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadFormatResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          }
        }
      },
      "get": {
        "summary": "Get all Locations",
        "description": "Get a list of all the Locations the current user has access to. Users with a free account can view Locations that they own and the one shared Location they are allowed to join. Premium users can view Locations that they own or are a member of. Locations can be searched by name and filtered by owner/member.",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "description": "Search term for searching Location by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "description": "Field to sort Locations by",
            "schema": {
              "type": "string",
              "default": "lastOpened",
              "enum": ["name", "createdAt", "updatedAt", "lastOpened"]
            }
          },
          {
            "$ref": "#/components/parameters/OwnerQuery"
          },
          {
            "$ref": "#/components/parameters/MemberQuery"
          },
          {
            "$ref": "#/components/parameters/OrderQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationBasic"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      }
    },
    "/locations/{id}": {
      "get": {
        "summary": "Get a Location by ID",
        "description": "Get a Location by ID. Will only return a Location if a user is authorized to see it. Users with a free account can view Locations that they own and the one shared Location they are allowed to join. Premium users can view Locations that they own or are a member of.",
        "parameters": [
          {
            "$ref": "#/components/parameters/LocationIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested Location object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationBasicValue"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      },
      "put": {
        "summary": "Update a Location",
        "description": "Update the Location with the given ID. Updatable fields include name, iconName, lastUpdated, notes, and notificationDays. Users with free accounts can update their default personal Location and their default shared Location, as long as the owner of the shared Location has an active subscription. Premium users can update Locations that they own or that they are a member of, provided that the owner of the Location has an active subscription",
        "parameters": [
          {
            "$ref": "#/components/parameters/LocationIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadFormatResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      },
      "delete": {
        "summary": "Delete a Location",
        "description": "Delete a Location by ID. Users can delete Locations that they own, regardless of subscription status.",
        "parameters": [
          {
            "$ref": "#/components/parameters/LocationIdParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LocationBasic": {
        "type": "object",
        "required": [
          "id",
          "name",
          "iconName",
          "owner",
          "notificationDays",
          "lastOpened",
          "items",
          "members",
          "lastUpdatedBy",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectID"
          },
          "name": {
            "$ref": "#/components/schemas/LocationName"
          },
          "iconName": {
            "$ref": "#/components/schemas/LocationIconName"
          },
          "owner": {
            "$ref": "#/components/schemas/BaseUserWithExpiry"
          },
          "members": {
            "type": "array",
            "description": "A list of members of the Location",
            "items": {
              "$ref": "#/components/schemas/BaseUser"
            }
          },
          "notificationDays": {
            "$ref": "#/components/schemas/LocationNotificationDays"
          },
          "items": {
            "type": "integer",
            "description": "The number of Items in the Location"
          },
          "notes": {
            "$ref": "#/components/schemas/LocationNotes"
          },
          "lastOpened": {
            "$ref": "#/components/schemas/LocationLastOpened"
          },
          "lastUpdatedBy": {
            "$ref": "#/components/schemas/BaseUser"
          },
          "createdAt": {
            "$ref": "#/components/schemas/CreatedAt"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UpdatedAt"
          }
        }
      },
      "LocationBasicValue": {
        "type": "object",
        "required": [
          "id",
          "name",
          "iconName",
          "owner",
          "notificationDays",
          "lastOpened",
          "items",
          "members",
          "totalValue",
          "lastUpdatedBy",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectID"
          },
          "name": {
            "$ref": "#/components/schemas/LocationName"
          },
          "iconName": {
            "$ref": "#/components/schemas/LocationIconName"
          },
          "owner": {
            "$ref": "#/components/schemas/BaseUserWithExpiry"
          },
          "members": {
            "type": "array",
            "description": "A list of members of the Location",
            "items": {
              "$ref": "#/components/schemas/BaseUser"
            }
          },
          "notificationDays": {
            "$ref": "#/components/schemas/LocationNotificationDays"
          },
          "items": {
            "type": "integer",
            "description": "The number of Items in the Location"
          },
          "totalValue": {
            "type": "number",
            "description": "The total value of the Items in this Location",
            "example": 101.99
          },
          "notes": {
            "$ref": "#/components/schemas/LocationNotes"
          },
          "lastOpened": {
            "$ref": "#/components/schemas/LocationLastOpened"
          },
          "lastUpdatedBy": {
            "$ref": "#/components/schemas/BaseUser"
          },
          "createdAt": {
            "$ref": "#/components/schemas/CreatedAt"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UpdatedAt"
          }
        }
      },
      "LocationCreateBody": {
        "type": "object",
        "required": ["name", "iconName"],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/LocationName"
          },
          "iconName": {
            "$ref": "#/components/schemas/LocationIconName"
          },
          "notes": {
            "$ref": "#/components/schemas/LocationNotes"
          }
        }
      },
...
      "LocationName": {
        "type": "string",
        "description": "The Location's name",
        "example": "Julia's Fridge",
        "maxLength": 100
      },
      "LocationIconName": {
        "type": "string",
        "description": "The Location's display icon name",
        "example": "fridge",
        "maxLength": 100
      },
...
    },
    "parameters": {
      "OwnerQuery": {
        "in": "query",
        "name": "owner",
        "description": "The id of the owner of this resource",
        "schema": {
          "$ref": "#/components/schemas/ObjectID"
        }
      },
      "MemberQuery": {
        "in": "query",
        "name": "member",
        "description": "The id of the member of this resource",
        "schema": {
          "$ref": "#/components/schemas/ObjectID"
        }
      },
      "LimitQuery": {
        "in": "query",
        "name": "limit",
        "required": false,
        "description": "The maximum number of results to return",
        "schema": {
          "type": "integer",
          "default": 30,
          "minimum": 0,
          "maximum": 50
        }
      },
...
      "OffsetQuery": {
        "in": "query",
        "name": "offset",
        "required": false,
        "description": "Offset of results to return",
        "schema": {
          "type": "integer",
          "default": "0",
          "minimum": 0
        }
      },
...
      "OrderQuery": {
        "in": "query",
        "name": "order",
        "required": false,
        "description": "Whether to sort in descending order",
        "schema": {
          "type": "string",
          "enum": ["asc", "desc"],
          "default": "asc"
        }
      },
      "LocationIdParam": {
        "in": "path",
        "name": "id",
        "required": true,
        "description": "The ID of the Location",
        "schema": {
          "$ref": "#/components/schemas/ObjectID"
        }
      },
...
    },
    "responses": {
      "SuccessResponse": {
        "description": "A success message",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["message"],
              "properties": {
                "message": {
                  "description": "A success message",
                  "type": "string",
                  "example": "Update successful"
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "The requested resource was not found or the user's account level does not have permission to access it.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnauthorizedResponse": {
        "description": "User is unauthorized. This can be due to invalid credentials or not having a premium account.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadFormatResponse": {
        "description": "The request was formed badly. Perhaps required fields are missing or invalid in the request body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}